Generalise QualificationsTableComponent
By moving it into the main namespace we can also use it in the supoort
support interface.
