+++ app/components/qualification_row_component.html.erb
